+++ packages/entities/entities-shared/src/components/entity-base-config-card/ConfigCardDisplay.cy.ts
@@ -32,15 +32,17 @@ describe('<ConfigCardDisplay />', () => {
       cy.get('.config-card-display-yaml').should('be.visible')
     })
 
-    it('renders the JSON content correctly', () => {
+    it('renders the JSON content and endpoint correctly', () => {
       cy.mount(ConfigCardDisplay, {
         props: {
+          fetcherUrl: 'https://cloud.konghq.com/us/gateway-manager/91e192e0-5981-4662-a37d-7b24272c9da3/routes/0af86198-9822-46e0-9028-47b173caf4aa',
           format: 'json',
           record,
         },
       })
 
       cy.get('.config-card-display-json').should('be.visible')
+      cy.get('.config-card-display-json-endpoint').should('be.visible')
     })
   })
 })