+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.9...@kong-ui-public/entities-plugins@2.4.10) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [2.4.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.8...@kong-ui-public/entities-plugins@2.4.9) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [2.4.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.7...@kong-ui-public/entities-plugins@2.4.8) (2023-12-12)
+
+
+### Bug Fixes
+
+* **useFetchUrlBuilder:** fix exact match filter for kong manager ([#999](https://github.com/Kong/public-ui-components/issues/999)) ([53070e3](https://github.com/Kong/public-ui-components/commit/53070e3940586c98924c4647af7ad31b08c43935))
+
+
+
+
+
 ## [2.4.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@2.4.6...@kong-ui-public/entities-plugins@2.4.7) (2023-12-12)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins