+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@2.1.8...@kong-ui-public/entities-shared@2.2.0) (2023-12-12)
+
+
+### Features
+
+* **entities-shared:** add the get endpoint in the config tables [KHCP-9930] ([#995](https://github.com/Kong/public-ui-components/issues/995)) ([08d709f](https://github.com/Kong/public-ui-components/commit/08d709f8cfed435cab6167c294c5e141be8b12b8))
+
+
+
+
+
+## [2.1.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@2.1.7...@kong-ui-public/entities-shared@2.1.8) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [2.1.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@2.1.6...@kong-ui-public/entities-shared@2.1.7) (2023-12-12)
+
+
+### Bug Fixes
+
+* **useFetchUrlBuilder:** fix exact match filter for kong manager ([#999](https://github.com/Kong/public-ui-components/issues/999)) ([53070e3](https://github.com/Kong/public-ui-components/commit/53070e3940586c98924c4647af7ad31b08c43935))
+
+
+
+
+
 ## [2.1.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@2.1.5...@kong-ui-public/entities-shared@2.1.6) (2023-12-12)