+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.1.12...@kong-ui-public/entities-upstreams-targets@2.1.13) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [2.1.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.1.11...@kong-ui-public/entities-upstreams-targets@2.1.12) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [2.1.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.1.10...@kong-ui-public/entities-upstreams-targets@2.1.11) (2023-12-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [2.1.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.1.9...@kong-ui-public/entities-upstreams-targets@2.1.10) (2023-12-12)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets